Update components.py
Added the '--no-downsize' argument to "put" in def getArgs(self, t) as it is available but not implemented. This is to prevent the default resizing of uploaded images to 20048 x 1536, and the compression of videos.

Without this, the non-destructive features kick in causing generation of copies renamed with 'NEWER' and 'OLDER' appended, on both local and remote.

Also I'm not fussed about over-writing files to get a sync, so in checkout() I've allowed an upload or download if there's a file diff. I've commented out the re-naming of files where they appended 'NEWER' and 'OLDER' to file names